Simplify AppConfig to only contain App Info.
Updated the app config removing the flavor/type and metadata
information so that the app config can be created automatically
by drivechain. The most useful thing in this module is the debugging
info and package name, and it adds a whole extra step to installation.
So this seems like a good compromise.